backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data collection
feedparser>=6.0.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.32.0
playwright>=1.44.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.9.0
certifi>=2024.0.0

# AI Processing (Milestone 2)
openai>=1.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # passlib is incompatible with bcrypt>=5.0
python-multipart>=0.0.9
jinja2>=3.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0
